+++ CHANGELOG.md
@@ -18,6 +18,8 @@
     - `truncate()` truncates long sentences at a given limit and append a suffix if needed
     - `singularize()` returns the singular form of a given word
     - `pluralize()` returns the plural form of a given word
+- Added the `iterating/` module, containing several helpers for iterables:
+    - `renumerate()`, a reverse enumerate implementation
 - Added the `i16g/` module, to help with locale management:
     - A helper class to dynamically load localization files from a package path
 - Added the `importing/` module, a collection of helpers for dynamic importing: